move image operations to seperate file
